Multi_Lang_Invoice_Extractor/requirements.txt
streamlit
google-generative-ai
python-dotenv
langchain
PyPDF2
chromadb